+++ src/com/ExpertMode/Modules/Wolf.java
@@ -5,7 +5,6 @@ import org.bukkit.Sound;
 import org.bukkit.SoundCategory;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.EntityType;
-import org.bukkit.entity.PigZombie;
 import org.bukkit.entity.Player;
 import org.bukkit.scheduler.BukkitRunnable;
 
@@ -15,54 +14,55 @@ import com.PluginBase.MathHelper;
 import com.PluginBase.SoundEmitter;
 
 /*
- * This class will randomly aggravate zombie pigmen towards the player
+ * This class will randomly aggravate wolfss towards the player
  */
-public class ZombiePigman extends Module {
+public class Wolf extends Module {
 
 	private final int 
-			aggravateZombiePigmenRange = 30,				// range from the player in which zombie pigmen will become aggresive
-			aggravationPeriod = 1200;						// amount of ticks between chance of aggravation
-	private final double aggravateZombiePigmenChange = 30;	// chance of zombie pigmen becoming aggresive
+			aggravateWolfRange = 15,					// range from the player in which wolfs will become aggresive
+			aggravationPeriod = 1200;					// amount of ticks between chance of aggravation
+	private final double aggravateWolfChange = 30;		// chance of wolfs becoming aggresive
 	
-	public ZombiePigman(Main main) {
+	public Wolf(Main main) {
 		super(main);
 		
 		// Start a timer that runs once every minute
 		BukkitRunnable bukkitRunnable = new BukkitRunnable() {
 			@Override
 			public void run() {
-				aggravateZombiePigmen();
+				aggravateWolfs();
 			}
 		};
 		bukkitRunnable.runTaskTimer(main, 0, this.aggravationPeriod);
 	}
 	
-	public void aggravateZombiePigmen() {
+	public void aggravateWolfs() {
 
 		// Loop through all the players
 		for (Player onlinePlayer : Bukkit.getOnlinePlayers()) {
 
 			// Check if the possibility hit
-			if (!MathHelper.getInstance().hasChanceHit(this.aggravateZombiePigmenChange)) {
+			if (!MathHelper.getInstance().hasChanceHit(this.aggravateWolfChange)) {
 				return;
 			}
+			
 			// Loop through nearby mobs
-			for (Entity nearbyEntity : onlinePlayer.getNearbyEntities(this.aggravateZombiePigmenRange, this.aggravateZombiePigmenRange, this.aggravateZombiePigmenRange)) {
+			for (Entity nearbyEntity : onlinePlayer.getNearbyEntities(this.aggravateWolfRange, this.aggravateWolfRange, this.aggravateWolfRange)) {
 
 				// Check if the entity is a zombie pigman
-				if (nearbyEntity.getType() == EntityType.PIG_ZOMBIE) {
+				if (nearbyEntity.getType() == EntityType.WOLF) {
 
-					// Get the zombie pigman
-					PigZombie pigZombie = (PigZombie) nearbyEntity;
-
-					// Make the zombie pigman attack the player
-					pigZombie.setTarget(onlinePlayer);
+					// Get the wolf
+					org.bukkit.entity.Wolf wolf = (org.bukkit.entity.Wolf) nearbyEntity;
+					
+					// Make the wolf attack the player
+					wolf.setTarget(onlinePlayer);
 					
 					// Play an angry sound at the location of the zombie pigman
 					SoundEmitter.getInstance().emitSound
 					(
 							onlinePlayer.getLocation().getWorld(),		// world
-							pigZombie.getLocation(),					// location
+							wolf.getLocation(),							// location
 							Sound.ENTITY_ZOMBIE_PIGMAN_ANGRY,			// sound
 							SoundCategory.HOSTILE,						// sound category
 							1,											// volume